engines/xp: cut redundant per-sample and per-tick computation
Profiling the SC-88 hot path found several cases of pure, verifiable
waste rather than any accuracy/performance trade-off - each verified
byte-identical WAV output against SC-88 and JV-880 before and after:

- oscillator.cc: value_at()/outer_tap() read 24-bit PCM samples into a
  double; a float's 24-bit mantissa represents every such sample
  exactly, so this is a width change with no precision loss. Also
  drop the std::floor() around the (always non-negative) phase
  accumulator - a plain truncating cast is the same integer, without
  std::floor's FPU-control-word churn on the 32-bit target.
- delay.cc/delay.h: the same exact-in-float argument for the stored
  delay tap positions, which are set once per control period rather
  than accumulated.
- tvf.cc: frequencyProgress() and the static-gain glide's curve
  decode were both re-deriving a result from a ROM word that never
  changes at runtime (the first is a hardcoded literal with three
  tests asserting it; the second is SC88_STATIC_AMPLITUDE_CURVE_WORD,
  a #define) on every call instead of once. tvf_word_to_hz's
  exp2+asin+sin chain is memoized per voice on the rounded frequency
  word, which changes far less often than it's read.
- engine.cc/renderer.h: hoisted a handful of per-voice values
  (period_fraction, the static-gain glide's exp() result) that are
  actually the same for every voice in a frame; bypassed a redundant
  C-linkage wrapper indirection in a C++-only call path; replaced a
  per-sample, per-voice, 3x-per-note ROM lookup for reverb/chorus/
  delay send gain with a 128-entry table computed once at renderer
  init, since the ROM never changes after that.

+++ libemusc/src/engines/xp/delay.cc
@@ -121,7 +121,7 @@ bool delay_set_params(const struct sc88_rom *rom, struct sc88_delay *dl,
     return false;
   unsigned centreUnits = centreWord - 0x8000u;
   double scale = dl->output_rate / (kDelayUnitsPerMs * 1000.0);
-  dl->centre_samples = (double)centreUnits * scale;
+  dl->centre_samples = (float)((double)centreUnits * scale);
 
   for (unsigned i = 0; i < 2; ++i) {
     uint8_t r = v[2 + i];
@@ -134,9 +134,9 @@ bool delay_set_params(const struct sc88_rom *rom, struct sc88_delay *dl,
         units = kDelayMaxUnits;
     }
     if (i == 0)
-      dl->left_samples = (double)units * scale;
+      dl->left_samples = (float)((double)units * scale);
     else
-      dl->right_samples = (double)units * scale;
+      dl->right_samples = (float)((double)units * scale);
   }
 
   /* `64*p` read as an XP coefficient is p/128 */
